fixed login code

+++ controllers/chatController.ts
@@ -9,6 +9,9 @@ export const New = async (req: Request, res: Response, next: NextFunction) => {
         const message = req.body.message;
         const account: Document<unknown, any, IAccount> & IAccount & { _id: Types.ObjectId; } = req.body.account;
 
+        if(!message)
+            return res.status(400).send({msg: config.err400 })
+
         // Remove old Chatbox
         if(account.chats.length > 0) {
             Chat.findByIdAndDelete(account.chats.pop()).exec((err, doc: any) => {
@@ -39,7 +42,7 @@ export const New = async (req: Request, res: Response, next: NextFunction) => {
             account.save()
             minChatSaler.chats.push(doc._id)
             minChatSaler.save()
-            return res.send({ msg: config.success })
+            return res.send({ msg: config.success, data: doc._id })
         })
     } catch (err) {
         console.log(err)
@@ -54,23 +57,24 @@ export const List = async (req: Request, res: Response, next: NextFunction) => {
 
 export const AddMessage = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const account: Document<unknown, any, IAccount> & IAccount & { _id: Types.ObjectId; } = req.body.account;
+        const account = req.body.account;
         const {_id, message} = req.body;
-        Chat.findById(_id, (err: Error, doc: Document<unknown, any, IChat> & IChat & { _id: Types.ObjectId; }) => {
-            if (err)
-                return res.status(500).send({msg: config.err500})
-            if(!doc)
-                return res.status(400).send({msg: config.err400})
-            if (!doc.customer.equals(account._id) && !doc.saler.equals(account._id))
-                return res.status(400).send({ msg: config.errPermission })
-            const isCustomer: boolean = doc.customer.equals(account._id);
-            //@ts-ignore
-            doc.messages.unshift({ isCustomer: isCustomer, message: message })
-            doc.save((_err) => {
-                if (_err)
-                    return res.status(500).send({msg: config.err500})
-                return res.send({ msg: config.success })
-            })
+        var find_options: any;
+        var search_options: any;
+
+        if(account.role == "Sale") {
+            find_options = {_id, saler: account._id}
+            search_options = { $push: { "messages": { $each: [{ isCustomer: false, message: message }], $position: 0 } } }
+        }
+        else {
+            find_options = {_id, customer: account._id}
+            search_options = { $push: { "messages": { $each: [{ isCustomer: true, message: message }], $position: 0 } } }
+        }
+
+        Chat.findOneAndUpdate(find_options, search_options).exec((err, doc) => {
+            if(err) return res.status(500).send({msg: config.err500})
+            if(!doc) return res.status(400).send({msg: config.err400})
+            return res.send({msg: config.success})
         })
     } catch (err) {
         console.log(err)
@@ -91,7 +95,7 @@ export const GetMessages = async (req: Request, res: Response, next: NextFunctio
                 return res.status(400).send({ msg: config.errPermission})
             
             //@ts-ignore
-            res.send({msg: config.success, doc: {_id: doc._id, customer: doc.customer.name, saler: doc.saler.name, messages: doc.messages}})
+            res.send({msg: config.success, doc: {_id: doc._id, customer: doc.customer.name, saler: doc.saler.name, messages: doc.messages.reverse()}})
         })
     } catch (err) {
         console.log(err)
